WIP move spawn manager
