+++ pgdrive/envs/marl_envs/marl_inout_roundabout.py
@@ -11,7 +11,6 @@ from pgdrive.scene_creator.blocks.first_block import FirstBlock
 from pgdrive.scene_creator.blocks.roundabout import Roundabout
 from pgdrive.scene_creator.map import PGMap
 from pgdrive.scene_creator.road.road import Road
-from pgdrive.scene_manager.spawn_manager import SpawnManager
 from pgdrive.utils import get_np_random, norm, PGConfig
 
 MARoundaboutConfig = dict(
@@ -135,23 +134,6 @@ class MultiAgentRoundaboutEnv(MultiAgentPGDrive):
             self.maps[self.current_seed] = new_map
             self.current_map = self.maps[self.current_seed]
 
-    def _process_extra_config(self, config):
-        config = super(MultiAgentRoundaboutEnv, self)._process_extra_config(config)
-        config = self._update_agent_pos_configs(config)
-        return super(MultiAgentRoundaboutEnv, self)._process_extra_config(config)
-
-    def _update_agent_pos_configs(self, config):
-        self._spawn_manager = SpawnManager(
-            spawn_roads=self.spawn_roads,
-            exit_length=config["map_config"]["exit_length"],
-            lane_num=config["map_config"]["lane_num"],
-            num_agents=config["num_agents"],
-            vehicle_config=config["vehicle_config"]
-        )
-        config["target_vehicle_configs"] = self._spawn_manager.get_target_vehicle_configs(
-            config["num_agents"], seed=self._DEBUG_RANDOM_SEED
-        )
-        return config
 
     def reset(self, *args, **kwargs):
         # Shuffle spawn places!
@@ -162,20 +144,20 @@ class MultiAgentRoundaboutEnv(MultiAgentPGDrive):
 
         # Multi-agent related reset
         # avoid create new observation!
-        obses = self.agent_manager.get_observations() or list(self.observations.values())
+        obses = self._agent_manager.get_observations() or list(self.observations.values())
         assert len(obses) == len(self.config["target_vehicle_configs"].keys())
         self.observations = {k: v for k, v in zip(self.config["target_vehicle_configs"].keys(), obses)}
         self.done_observations = dict()
 
         # Must change in-place!
-        obs_spaces = self.agent_manager.get_observation_spaces() or list(
+        obs_spaces = self._agent_manager.get_observation_spaces() or list(
             self.observation_space.spaces.values()
         )
         assert len(obs_spaces) == len(self.config["target_vehicle_configs"].keys())
         for o in obs_spaces:
             assert isinstance(o, Box)
         self.observation_space.spaces = {k: v for k, v in zip(self.observations.keys(), obs_spaces)}
-        action_spaces = self.agent_manager.get_action_spaces() or list(self.action_space.spaces.values())
+        action_spaces = self._agent_manager.get_action_spaces() or list(self.action_space.spaces.values())
         self.action_space.spaces = {k: v for k, v in zip(self.observations.keys(), action_spaces)}
 
         ret = PGDriveEnvV2.reset(self, *args, **kwargs)
@@ -183,39 +165,13 @@ class MultiAgentRoundaboutEnv(MultiAgentPGDrive):
         assert len(self.vehicles) == self.num_agents
         self.for_each_vehicle(self._update_destination_for)
 
-        self.agent_manager.reset(
+        self._agent_manager.reset(
             vehicles=self.vehicles,
             observation_spaces=self.observation_space.spaces,
             observations=self.observations,
             action_spaces=self.action_space.spaces
         )
         return ret
-
-    def step(self, actions):
-        o, r, d, i = super(MultiAgentRoundaboutEnv, self).step(actions)
-
-        # Update respawn manager
-        if self.episode_steps >= self.config["horizon"]:
-            self.agent_manager.set_allow_respawn(False)
-        self._spawn_manager.update(self.vehicles, self.current_map)
-        new_obs_dict = self._respawn()
-        if new_obs_dict:
-            for new_id, new_obs in new_obs_dict.items():
-                o[new_id] = new_obs
-                r[new_id] = 0.0
-                i[new_id] = {}
-                d[new_id] = False
-
-        # Update __all__
-        d["__all__"] = (
-                ((self.episode_steps >= self.config["horizon"]) and (all(d.values()))) or (len(self.vehicles) == 0)
-                or (self.episode_steps >= 5 * self.config["horizon"])
-        )
-        if d["__all__"]:
-            for k in d.keys():
-                d[k] = True
-
-        return o, r, d, i
 
     def _update_destination_for(self, vehicle):
         # when agent re-joined to the game, call this to set the new route to destination
@@ -236,7 +192,7 @@ class MultiAgentRoundaboutEnv(MultiAgentPGDrive):
         """
         This function can force a given vehicle to respawn!
         """
-        self.agent_manager.finish(agent_name)
+        self._agent_manager.finish(agent_name)
         new_id, new_obs = self._respawn_single_vehicle()
         return new_id, new_obs
 
@@ -244,7 +200,7 @@ class MultiAgentRoundaboutEnv(MultiAgentPGDrive):
         """
         Arbitrary insert a new vehicle to a new spawn place if possible.
         """
-        allow_respawn, vehicle_info = self.agent_manager.propose_new_vehicle()
+        allow_respawn, vehicle_info = self._agent_manager.propose_new_vehicle()
         if vehicle_info is None:  # No more vehicle to be assigned.
             return None, None
         if not allow_respawn:
@@ -252,16 +208,16 @@ class MultiAgentRoundaboutEnv(MultiAgentPGDrive):
             v = vehicle_info["vehicle"]
             v.set_position((-999, -999))
             v.set_static(True)
-            self.agent_manager.confirm_respawn(False, vehicle_info)
+            self._agent_manager.confirm_respawn(False, vehicle_info)
             return None, None
         v = vehicle_info["vehicle"]
         dead_vehicle_id = vehicle_info["old_name"]
         bp_index = self._replace_vehicles(v)
         if bp_index is None:  # No more spawn places to be assigned.
-            self.agent_manager.confirm_respawn(False, vehicle_info)
+            self._agent_manager.confirm_respawn(False, vehicle_info)
             return None, None
 
-        self.agent_manager.confirm_respawn(True, vehicle_info)
+        self._agent_manager.confirm_respawn(True, vehicle_info)
 
         new_id = vehicle_info["new_name"]
         v.set_static(False)
@@ -286,13 +242,13 @@ class MultiAgentRoundaboutEnv(MultiAgentPGDrive):
                     self.dones[v_id] = True
         for dead_vehicle_id, done in dones.items():
             if done:
-                self.agent_manager.finish(dead_vehicle_id)
+                self._agent_manager.finish(dead_vehicle_id)
                 self.vehicles.pop(dead_vehicle_id)
                 self.action_space.spaces.pop(dead_vehicle_id)
         return obs, reward, dones, info
 
     def _reset_vehicles(self):
-        vehicles = self.agent_manager.get_vehicle_list() or list(self.vehicles.values())
+        vehicles = self._agent_manager.get_vehicle_list() or list(self.vehicles.values())
         assert len(vehicles) == len(self.observations)
         self.vehicles = {k: v for k, v in zip(self.observations.keys(), vehicles)}
         self.done_vehicles = {}
@@ -370,7 +326,7 @@ def _expert():
         if d["__all__"]:
             print(
                 "Finish! Current step {}. Group Reward: {}. Average reward: {}".format(
-                    i, total_r, total_r / env.agent_manager.next_agent_count
+                    i, total_r, total_r / env._agent_manager.next_agent_count
                 )
             )
             break
@@ -420,7 +376,7 @@ def _vis():
         if d["__all__"]:
             print(
                 "Finish! Current step {}. Group Reward: {}. Average reward: {}".format(
-                    i, total_r, total_r / env.agent_manager.next_agent_count
+                    i, total_r, total_r / env._agent_manager.next_agent_count
                 )
             )
             # break
